Move the v2 WebRTC client into web resources
Keep browser markup and behavior independently editable while packaging and serving both assets with the runtime.
