+++ flashdreams/flashdreams/runtime_v2/serving/web/index.html
@@ -1,0 +1,18 @@
+<!doctype html>
+<!--
+SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+SPDX-License-Identifier: Apache-2.0
+-->
+
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <title>FlashDreams WebRTC</title>
+  </head>
+  <body>
+    <video id="video" autoplay playsinline></video>
+    <button id="activate" type="button">Activate</button>
+    <button id="reset" type="button">Reset</button>
+    <script src="/app.js"></script>
+  </body>
+</html>
